Act on collection ToDos to integrate with scala.collection namespace
We can now move new methods into non-collection classes like `Array`
and access `private[scala]` members (e.g. in `MurmurHash3`).
