+++ src/library/scala/collection/immutable/IntMap.scala
@@ -390,17 +390,17 @@ sealed abstract class IntMap[+T] extends Map[Int, T]
   def unionWith[S >: T](that: IntMap[S], f: (Int, S, S) => S): IntMap[S] = (this, that) match{
     case (IntMap.Bin(p1, m1, l1, r1), that@(IntMap.Bin(p2, m2, l2, r2))) =>
       if (shorter(m1, m2)) {
-        if (!hasMatch(p2, p1, m1)) join[S](p1, this, p2, that) // TODO: remove [S] when scala/bug#5548 is fixed
+        if (!hasMatch(p2, p1, m1)) join(p1, this, p2, that)
         else if (zero(p2, m1)) IntMap.Bin(p1, m1, l1.unionWith(that, f), r1)
         else IntMap.Bin(p1, m1, l1, r1.unionWith(that, f))
       } else if (shorter(m2, m1)){
-        if (!hasMatch(p1, p2, m2)) join[S](p1, this, p2, that) // TODO: remove [S] when scala/bug#5548 is fixed
+        if (!hasMatch(p1, p2, m2)) join(p1, this, p2, that)
         else if (zero(p1, m2)) IntMap.Bin(p2, m2, this.unionWith(l2, f), r2)
         else IntMap.Bin(p2, m2, l2, this.unionWith(r2, f))
       }
       else {
         if (p1 == p2) IntMap.Bin(p1, m1, l1.unionWith(l2,f), r1.unionWith(r2, f))
-        else join[S](p1, this, p2, that) // TODO: remove [S] when scala/bug#5548 is fixed
+        else join(p1, this, p2, that)
       }
     case (IntMap.Tip(key, value), x) => x.updateWith[S](key, value, (x, y) => f(key, y, x))
     case (x, IntMap.Tip(key, value)) => x.updateWith[S](key, value, (x, y) => f(key, x, y))
