+++ src/library/scala/collection/mutable/AnyRefMap.scala
@@ -422,16 +422,15 @@ class AnyRefMap[K <: AnyRef, V] private[collection] (defaultEntry: K => V, initi
   def clear(): Unit = keysIterator() foreach -=
 
   // The `K with AnyRef` parameter type is necessary to distinguish these methods from the base methods they overload (not override)
-  // TODO: Remove the unnecessary implicit in Scala 2.13; Dotty requires it for disambiguation
-  def map[K2 <: AnyRef, V2](f: ((K with AnyRef, V)) => (K2, V2))(implicit ev: K2 <:< AnyRef): AnyRefMap[K2, V2] =
+  def map[K2 <: AnyRef, V2](f: ((K with AnyRef, V)) => (K2, V2)): AnyRefMap[K2, V2] =
     AnyRefMap.from(new View.Map(toIterable, f))
-  def flatMap[K2 <: AnyRef, V2](f: ((K with AnyRef, V)) => IterableOnce[(K2, V2)])(implicit ev: K2 <:< AnyRef): AnyRefMap[K2, V2] =
+  def flatMap[K2 <: AnyRef, V2](f: ((K with AnyRef, V)) => IterableOnce[(K2, V2)]): AnyRefMap[K2, V2] =
     AnyRefMap.from(new View.FlatMap(toIterable, f))
-  def collect[K2 <: AnyRef, V2](pf: PartialFunction[(K with AnyRef, V), (K2, V2)])(implicit ev: K2 <:< AnyRef): AnyRefMap[K2, V2] =
+  def collect[K2 <: AnyRef, V2](pf: PartialFunction[(K with AnyRef, V), (K2, V2)]): AnyRefMap[K2, V2] =
     flatMap { kv: (K with AnyRef, V) =>
       if (pf.isDefinedAt(kv)) new View.Single(pf(kv))
       else View.Empty
-    }(ev)
+    }
 }
 
 object AnyRefMap {
