+++ src/library/scala/collection/mutable/WrappedArray.scala
@@ -131,7 +131,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): T = array(index).asInstanceOf[T]
     def update(index: Int, elem: T): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofRef[_] => Arrays.equals(array.asInstanceOf[Array[AnyRef]], that.array.asInstanceOf[Array[AnyRef]])
       case _ => super.equals(that)
@@ -144,7 +144,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Byte = array(index)
     def update(index: Int, elem: Byte): Unit = { array(index) = elem }
-    override def hashCode = wrappedBytesHash(array)
+    override def hashCode = MurmurHash3.wrappedBytesHash(array)
     override def equals(that: Any) = that match {
       case that: ofByte => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -157,7 +157,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Short = array(index)
     def update(index: Int, elem: Short): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofShort => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -170,7 +170,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Char = array(index)
     def update(index: Int, elem: Char): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofChar => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -183,7 +183,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Int = array(index)
     def update(index: Int, elem: Int): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofInt => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -196,7 +196,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Long = array(index)
     def update(index: Int, elem: Long): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofLong => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -209,7 +209,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Float = array(index)
     def update(index: Int, elem: Float): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofFloat => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -222,7 +222,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Double = array(index)
     def update(index: Int, elem: Double): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofDouble => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -235,7 +235,7 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Boolean = array(index)
     def update(index: Int, elem: Boolean): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofBoolean => Arrays.equals(array, that.array)
       case _ => super.equals(that)
@@ -248,14 +248,10 @@ object WrappedArray extends StrictOptimizedClassTagSeqFactory[WrappedArray] { se
     def length: Int = array.length
     def apply(index: Int): Unit = array(index)
     def update(index: Int, elem: Unit): Unit = { array(index) = elem }
-    override def hashCode = wrappedArrayHash(array)
+    override def hashCode = MurmurHash3.wrappedArrayHash(array)
     override def equals(that: Any) = that match {
       case that: ofUnit => array.length == that.array.length
       case _ => super.equals(that)
     }
   }
-
-  //TODO Use MurmurHash3.wrappedArrayHash/wrappedBytesHash which are private[scala]
-  private def wrappedArrayHash[@specialized T](a: Array[T]): Int  = MurmurHash3.arrayHash(a, MurmurHash3.seqSeed)
-  private def wrappedBytesHash(data: Array[Byte]): Int            = MurmurHash3.bytesHash(data, MurmurHash3.seqSeed)
 }
